requirements.txt
# LLM & Embedding
langchain>=1.0.0
langchain-core>=1.0.0
langchain-google-genai>=2.0.0
langchain-chroma>=0.1.0
google-generativeai>=0.8.0

# Vector DB
chromadb>=0.5.0

# LangGraph (Agent)
langgraph>=1.0.0
langgraph-prebuilt>=1.0.0

# Document Parsing
python-docx>=1.1.0
pdfplumber>=0.10.0
pypdf>=4.0.0
openpyxl>=3.1.0

# UI
gradio>=4.0.0

# Utilities
python-dotenv>=1.0.0
tqdm>=4.66.0
